Also style the editor page
