+++ gatsby/src/components/avatar.js
@@ -1,0 +1,21 @@
+import * as React from "react";
+import { Link } from "gatsby";
+import { GatsbyImage, getImage } from "gatsby-plugin-image";
+
+import * as styles from "./avatar.module.css";
+
+const Avatar = ({ author, rounded }) => {
+  const imageClassName = (rounded) ? styles.avatarRounded : styles.avatar;
+
+  const image = getImage(author.avatar.localFile);
+  return (
+    <Link to={`/editors/${author.user.username}`}>
+      <GatsbyImage
+        className={imageClassName}
+        image={image}
+        alt={"avatar of " + author.fullName} />
+    </Link>
+  );
+}
+
+export default Avatar;
