+++ gatsby/src/components/avatar.module.css
@@ -1,0 +1,27 @@
+.avatar {
+  justify-self: center;
+
+  max-width: 100%;
+  height: auto;
+
+  box-shadow:
+  1px 0.5px 1.9px -1px rgba(0, 0, 0, 0.6),
+  1.4px 0.7px 3.7px -1px rgba(0, 0, 0, 0.9),
+  1.6px 0.8px 6.4px -1px rgba(250, 218, 173, 0.6),
+  2px 1px 14px -1px rgba(250, 218, 173, 0.9)
+}
+
+.avatarRounded {
+  justify-self: center;
+
+  border-radius: 50%;
+  max-width: 100%;
+  height: auto;
+
+  box-shadow:
+  1px 0.5px 1.9px -1px rgba(0, 0, 0, 0.6),
+  1.4px 0.7px 3.7px -1px rgba(0, 0, 0, 0.9),
+  1.6px 0.8px 6.4px -1px rgba(250, 218, 173, 0.6),
+  2px 1px 14px -1px rgba(250, 218, 173, 0.9)
+}
+
