+++ gatsby/src/components/layout/layout.js
@@ -7,11 +7,11 @@ import * as styles from "./layout.module.css";
 const Layout = ({ title, children }) => {
   return (
     <>
-    <Navbar title={title} linkTo="/" />
-    <div className={styles.wrapper}>
-      {children}
-      <footer className={styles.footer}>© {new Date().getFullYear()}, {title}</footer>
-    </div>
+      <Navbar title={title} linkTo="/" />
+      <div className={styles.wrapper}>
+        {children}
+        <footer className={styles.footer}>© {new Date().getFullYear()}, {title}</footer>
+      </div>
     </>
   );
 };
