+++ gatsby/src/components/post/info.js
@@ -1,21 +1,11 @@
 import * as React from "react";
 import { Link } from "gatsby";
-import { GatsbyImage, getImage } from "gatsby-plugin-image";
+
+import Avatar from "../avatar";
 
 import * as styles from "./info.module.css";
 
 const PostInfo = ({ date, author }) => {
-  const image = getImage(author.avatar.localFile);
-  const avatarBlock = (
-    <Link to={`/editors/${author.user.username}`}>
-      <GatsbyImage
-        className={styles.avatar}
-        image={image}
-        alt={"avatar of " + author.fullName} />
-    </Link>
-  );
-
-
   const authorDateBlock = (
     <div className={styles.author}>
       <Link className={styles.authorName} to={`/editors/${author.user.username}`}>
@@ -27,7 +17,7 @@ const PostInfo = ({ date, author }) => {
 
   return (
     <div className={styles.info}>
-      { avatarBlock }
+      <Avatar author={author} />
       { authorDateBlock }
     </div>
   );
