+++ gatsby/src/components/post/info.module.css
@@ -1,11 +1,3 @@
-.avatar {
-  justify-self: center;
-
-  border-radius: 50%;
-  max-width: 100%;
-  height: auto;
-}
-
 .info {
   height: 4rem;
