+++ gatsby/src/pages/editors/{StrapiEditor.user__username}.js
@@ -1,33 +1,16 @@
 import * as React from "react";
-import { graphql, Link } from "gatsby";
-import { GatsbyImage, getImage } from "gatsby-plugin-image";
+import { graphql } from "gatsby";
 
 import Layout from "../../components/layout/layout";
+import Editor from "../../components/editor";
 
 function EditorPage({ data }) {
   const { strapiEditor, site } = data;
   const siteTitle = site.siteMetadata?.title || `Title`;
 
-  const { avatar, bio, fullName, posts } = strapiEditor;
-  const image = getImage(avatar.localFile);
-
   return (
     <Layout title={siteTitle}>
-      <h1>{ fullName }</h1>
-      <GatsbyImage
-        image={image}
-        alt={"avatar of " + fullName} />
-      <h2>Bio</h2>
-      <div dangerouslySetInnerHTML={{ __html: bio.data.childMarkdownRemark.html }}/>
-      <h2>Posts</h2>
-      <ul>
-        {
-          posts.map(post => <li>
-          <Link to={"/posts/" + post.slug}>{ post.title }</Link>
-          </li>)
-        }
-      </ul>
-
+      <Editor editor={strapiEditor} />
     </Layout>
   );
 };
@@ -56,7 +39,7 @@ export const pageQuery = graphql`
       avatar {
         localFile {
           childImageSharp {
-            gatsbyImageData(width: 300)
+            gatsbyImageData(width: 200)
           }
         }
       }
@@ -66,6 +49,7 @@ export const pageQuery = graphql`
       posts {
         title
         slug
+        date
       }
     }
   }
